requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
albumentations==0.4.3
altair==4.2.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
braceexpand==0.1.7
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
-e git+https://github.com/openai/CLIP.git@main#egg=clip
contourpy==1.2.1
cycler==0.12.1
datasets==2.20.0
dill==0.3.8
dnspython==2.6.1
docker-pycreds==0.4.0
drjit==0.4.6
einops==0.3.0
email_validator==2.2.0
entrypoints==0.4
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
ftfy==6.2.0
future==1.0.0
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.30.0
google-auth-oauthlib==0.4.6
gradio==3.42.0
gradio_client==0.5.0
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib_metadata==8.0.0
importlib_resources==6.4.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
kornia==0.7.2
kornia_rs==0.1.4
lazy_loader==0.4
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
mitsuba==3.5.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
nltk==3.8.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.40
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.1
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
platformdirs==4.2.2
protobuf==3.20.1
psutil==6.0.0
pudb==2019.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
pyDeprecate==0.3.1
pydub==0.25.1
Pygments==2.18.0
Pympler==1.0.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==1.4.2
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.20.0
scikit-learn==1.5.0
scipy==1.14.0
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.7.1
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.22.0
sympy==1.12.1
-e git+https://github.com/CompVis/taming-transformers.git@master#egg=taming-transformers
tenacity==8.4.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
test_tube==0.7.5
threadpoolctl==3.5.0
tifffile==2024.6.18
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torch-fidelity==0.3.0
torchmetrics==0.6.0
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.3
triton==2.3.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.2.2
urwid==2.6.14
uvicorn==0.30.1
uvloop==0.19.0
validators==0.28.3
wandb==0.17.3
watchdog==4.0.1
watchfiles==0.22.0
wcwidth==0.2.13
webdataset==0.2.5
websockets==11.0.3
Werkzeug==3.0.3
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
